backend/requirements.txt
asgiref==3.11.0
atomicwrites==1.4.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
Django==6.0
defusedxml==0.7.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
flake8==7.3.0
flake8-isort==7.0.0
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
py==1.11.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-dotenv==1.2.1
python3-openid==3.2.0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
toml==0.10.2
tzdata==2025.3
urllib3==2.6.2
